requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
django-debug-toolbar==4.1.0
django-djconfig==0.11.0
django-haystack==3.2.1
django-infinite-scroll-pagination==1.1.0
django-js-asset==2.0.0
django-machina==1.3.0
django-mptt==0.14.0
django-widget-tweaks==1.4.12
idna==3.4
markdown2==2.4.8
mistune==0.8.4
olefile==0.46
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
Whoosh==2.7.4
gunicorn
psycopg2-binary
